Init Commit
So far this supports viewing and navigating images in a folder. It can autoscale and zoom to full width. Moving around the image while zoomed is still very janky. The code for the Config class is neat but also janky. Gonna break this up into seperate files in a future commit.